Moved storage writer open and close out of analysis engine
